mod：多人连麦，和人机的开口时长转成int再提交

+++ src/main/java/com/xueersi/parentsmeeting/modules/livevideo/business/RolePlayerBll.java
@@ -735,7 +735,7 @@ public class RolePlayerBll extends BaseBll implements RolePlayAction {
                     JSONObject objAn = new JSONObject();
                     objAn.put("sentenceNum", i);
                     //上传的是自己朗读的有效时长
-                    objAn.put("entranceTime", message.getSelfValidSpeechTime());
+                    objAn.put("entranceTime", (int)message.getSelfValidSpeechTime());
                     objAn.put("score", message.getSpeechScore());
                     JSONObject objData = new JSONObject();
                     objData.put("cont_score", message.getFluency());
@@ -821,7 +821,7 @@ public class RolePlayerBll extends BaseBll implements RolePlayAction {
                 if (message.getRolePlayer().isSelfRole()) {
                     JSONObject objAn = new JSONObject();
                     objAn.put("sentenceNum", i);
-                    objAn.put("entranceTime", message.getSelfValidSpeechTime());
+                    objAn.put("entranceTime", (int)message.getSelfValidSpeechTime());
                     objAn.put("score", message.getSpeechScore());
                     JSONObject objData = new JSONObject();
                     objData.put("cont_score", message.getFluency());
